refactor(helpers): move utilities to the separate folder
