+++ src/control/controlApp.js
@@ -2,7 +2,7 @@ import { createApp } from 'stapp'
 import { persist, toAsync } from 'stapp/lib/modules/persist'
 import { stopWatch } from './modules/stopWatch/module'
 import { presentationState } from './modules/presentation/module'
-import { createConsumer } from './utils/createConsumer'
+import { createConsumer } from '../helpers/preact/createConsumer'
 import { config } from './modules/config/module'
 
 const app = createApp({
